# frontend/src-gen/api-docs.json
{
  "openapi" : "3.1.0",
  "info" : {
    "title" : "CampusCoffeeConsumption API",
    "description" : "REST API for tracking the coffee consumption of SE@UHD users: users bump their own count via a secret capability link, and admins manage users and adjust counts.",
    "version" : "0.0.0"
  },
...
  "paths" : {
...
    "/api/beans/{id}" : {
      "put" : {
        "tags" : [ "Beans" ],
        "summary" : "Rename a coffee bean (admin only).",
        "operationId" : "rename",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BeanRenameDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoffeeBeanDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/beans/{id}/merge" : {
      "post" : {
        "tags" : [ "Beans" ],
        "summary" : "Merge a coffee bean into another (admin only).",
        "operationId" : "merge",
        "parameters" : [ {
          "name" : "id",
          "in" : "path",
          "required" : true,
          "schema" : {
            "type" : "string",
            "format" : "uuid"
          }
        } ],
        "requestBody" : {
          "content" : {
            "application/json" : {
              "schema" : {
                "$ref" : "#/components/schemas/BeanMergeDto"
              }
            }
          },
          "required" : true
        },
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/CoffeeBeanDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/summary" : {
      "get" : {
        "tags" : [ "Self-service" ],
        "summary" : "Get the authenticated user's landing summary.",
        "operationId" : "summary",
        "parameters" : [ {
          "name" : "limit",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "exclusiveMinimum" : 0,
            "maximum" : 100
          }
        }, {
          "name" : "offset",
          "in" : "query",
          "required" : false,
          "schema" : {
            "type" : "integer",
            "format" : "int32",
            "maximum" : 1000000,
            "minimum" : 0
          }
        } ],
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "$ref" : "#/components/schemas/UserSummaryDto"
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
    "/api/beans" : {
      "get" : {
        "tags" : [ "Beans" ],
        "summary" : "List the selectable coffee beans.",
        "operationId" : "list_1",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CoffeeBeanDto"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
    "/api/beans/ratings" : {
      "get" : {
        "tags" : [ "Beans" ],
        "summary" : "List the bean ratings.",
        "operationId" : "ratings",
        "responses" : {
          "200" : {
            "description" : "OK",
            "content" : {
              "*/*" : {
                "schema" : {
                  "type" : "array",
                  "items" : {
                    "$ref" : "#/components/schemas/CoffeeBeanRatingsDto"
                  }
                }
              }
            }
          },
          "400" : {
            "description" : "The request is malformed or violates a validation rule.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "401" : {
            "description" : "The request is unauthenticated (a missing, unknown, or rotated credential).",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "403" : {
            "description" : "The authenticated caller is not permitted to perform this operation.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "404" : {
            "description" : "A referenced resource does not exist.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          },
          "409" : {
            "description" : "The request conflicts with the current state of the resource.",
            "content" : {
              "application/json" : {
                "schema" : {
                  "$ref" : "#/components/schemas/ErrorResponse"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components" : {
    "schemas" : {
...
      "ActivityEntryDto" : {
        "type" : "object",
        "properties" : {
          "type" : {
            "type" : "string",
            "enum" : [ "CONSUMPTION", "CONSUMPTION_CANCEL", "PRIVATE_EXPENSE", "DEPOSIT", "KITTY_EXPENSE", "KITTY_ADJUSTMENT", "PRICE_CHANGE" ]
          },
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "createdAt" : {
            "type" : "string",
            "format" : "date-time"
          },
          "createdBy" : {
            "type" : "string"
          },
          "note" : {
            "type" : [ "string", "null" ]
          },
          "amountCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "runningBalanceCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "count" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "delta" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "weightGrams" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          },
          "privateAmountCents" : {
            "type" : [ "integer", "null" ],
            "format" : "int64"
          },
          "kittyAmountCents" : {
            "type" : [ "integer", "null" ],
            "format" : "int64"
          }
        },
        "required" : [ "amountCents", "createdAt", "createdBy", "id", "runningBalanceCents", "type" ]
      },
      "CoffeeRatingPromptDto" : {
        "type" : "object",
        "properties" : {
          "canRate" : {
            "type" : "boolean"
          },
          "defaultBeanId" : {
            "type" : [ "string", "null" ],
            "format" : "uuid"
          },
          "value" : {
            "type" : [ "integer", "null" ],
            "format" : "int32"
          }
        },
        "required" : [ "canRate" ]
      },
      "UserSummaryDto" : {
        "type" : "object",
        "properties" : {
          "count" : {
            "type" : "integer",
            "format" : "int32"
          },
          "priceCents" : {
            "type" : "integer",
            "format" : "int32"
          },
          "balanceCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "kittyBalanceCents" : {
            "type" : "integer",
            "format" : "int64"
          },
          "cancellable" : {
            "type" : "boolean"
          },
          "summaryPanel" : {
            "type" : "string",
            "enum" : [ "BALANCE", "CUPS" ]
          },
          "firstCupAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "cupsThisWeek" : {
            "type" : "integer",
            "format" : "int32"
          },
          "cupsToday" : {
            "type" : "integer",
            "format" : "int32"
          },
          "ratingPrompt" : {
            "$ref" : "#/components/schemas/CoffeeRatingPromptDto"
          },
          "activity" : {
            "type" : "array",
            "items" : {
              "$ref" : "#/components/schemas/ActivityEntryDto"
            }
          }
        },
        "required" : [ "activity", "balanceCents", "cancellable", "count", "cupsThisWeek", "cupsToday", "kittyBalanceCents", "priceCents", "ratingPrompt", "summaryPanel" ]
      },
      "BeanRenameDto" : {
        "type" : "object",
        "properties" : {
          "name" : {
            "type" : "string",
            "maxLength" : 200,
            "minLength" : 0
          }
        },
        "required" : [ "name" ]
      },
      "CoffeeBeanDto" : {
        "type" : "object",
        "properties" : {
          "id" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "active" : {
            "type" : "boolean"
          }
        },
        "required" : [ "active", "id", "name" ]
      },
...
      "BeanMergeDto" : {
        "type" : "object",
        "properties" : {
          "targetBeanId" : {
            "type" : "string",
            "format" : "uuid"
          }
        },
        "required" : [ "targetBeanId" ]
      },
...
      "CoffeeBeanRatingsDto" : {
        "type" : "object",
        "properties" : {
          "beanId" : {
            "type" : "string",
            "format" : "uuid"
          },
          "name" : {
            "type" : "string"
          },
          "averageValue" : {
            "type" : [ "number", "null" ],
            "format" : "double"
          },
          "voteCount" : {
            "type" : "integer",
            "format" : "int32"
          },
          "latestRatingAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          },
          "latestPurchaseAt" : {
            "type" : [ "string", "null" ],
            "format" : "date-time"
          }
        },
        "required" : [ "beanId", "name", "voteCount" ]
      },
...
      "ErrorResponse" : {
        "properties" : {
          "errorCode" : {
            "type" : "string"
          },
          "message" : {
            "type" : "string"
          },
          "statusCode" : {
            "type" : "integer",
            "format" : "int32"
          },
          "statusMessage" : {
            "type" : "string"
          },
          "timestamp" : {
            "type" : "string",
            "format" : "date-time"
          },
          "path" : {
            "type" : "string"
          }
        }
      }
    }
  }
}